Reviewed stat_analysis notebook and created Modelling notebook
This notebook was created to work on the final modelling task (predicting the cgpa_change effect of the strike). The previous branch was outdated and needed a pull request before it could accept any change from origin. To avoid merge clashes and other errors, I stashed that branch and created this.
My changes are as follows:
1. I fixed some minor display errors in the cleaning notebook.
2. The newly cleaning tasks carried out caused some modifications to the data so I reviewed the result of the stat_analysis notebook.
        a. The affected features (faculties and course_units) changed in magnitude but remained unchanged in direction.
        b. I changed the range of the course units split from (8-15, 16-24) to (8-17, 18-24). This was done because of the unfairly low representation of the (8-15) class in the previous split. This took the p-value from 0.435 to 0.960 further highlighting that there is no significant difference between the two groups.

3. Modelling Notebook:
a. I created a new column called 'prep_change' that combines the 'prep_after' and 'prep_before'.
b. I grouped the 'cgpa_change' column (target variable) into three categories as the column 'cgpa_change_cat'.
c. I resampled the data to achieve class balance.
d. I tested 6 different models to see which performs best and I settled on the RandomForestClassifier (mostly this) and the LGBMClassifier.
e. I won't list all the preprocessing steps I took but they were largely model manipulation steps (Trying to decide which models and model configurations work best).
f. The best model I achieved has an accuracy of  0.640 with a precision of 75 for class 0, 53.3 for class 1 and 60.98 for class 2.
g. I will proceed to try and improve performance through preprocessing.